- added radar source icons for candi and foursquare. - sources with multiple links now rollup. - link picker. - start of classes to support service notifications. - action bar icons for link picker and candi form. - concurrency improvements. - removed source.icon and source.hidden. - added source.photo - removed entity.enabled (used by service only). - removed contact.twitter property. - serialization supports object and property level updates/inserts. - bug fixes.

+++ src/com/aircandi/ui/user/CandiUser.java
@@ -159,7 +159,7 @@ public class CandiUser extends CandiActivity {
 	public void onImageClick(View view) {
 		Intent intent = null;
 		final Photo photo = mUser.photo;
-		photo.setCreatedAt(mUser.photo.getCreatedAt());
+		photo.setCreatedAt(mUser.photo.getCreatedAt().longValue());
 		photo.setTitle(mUser.name);
 		photo.setUser(mUser);
 		ProxiManager.getInstance().getEntityModel().getPhotos().clear();
